Move consensus unit tests to separate file
Now we have split the `encode` module up; move the `consenus::encode`
unit tests to a separate file.
